+++ sbin/ifconfig/ifcarp.c
@@ -32,21 +32,18 @@
 #include <sys/ioctl.h>
 #include <sys/socket.h>
 #include <sys/sockio.h>
-
-#include <stdlib.h>
-#include <unistd.h>
-
-#include <net/ethernet.h>
 #include <net/if.h>
+#include <net/route.h>
+#include <net/ethernet.h>
 #include <netinet/in.h>
 #include <netinet/ip_carp.h>
-#include <net/route.h>
-
 #include <arpa/inet.h>
 
 #include <ctype.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
+#include <unistd.h>
 #include <err.h>
 #include <errno.h>
 
@@ -250,7 +247,7 @@ static struct afswtch af_carp = {
 static __constructor(101) void
 carp_ctor(void)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < nitems(carp_cmds);  i++)
 		cmd_register(&carp_cmds[i]);
